requirements.txt
streamlit
keras
mtcnn
python-opencv
Pillow